grouped QTargetBeat and QTargetMeasure into qtargetitems
