+++ abjadext/nauert/qtargetitems.py
@@ -2,8 +2,284 @@ import typing
 
 import abjad
 
-from .QTargetBeat import QTargetBeat
+from .qeventproxy import QEventProxy
+from .qgrid import QGrid
+from .QuantizationJob import QuantizationJob
+from .qevents import QEvent
 from .searchtrees import SearchTree, UnweightedSearchTree
+
+
+class QTargetBeat:
+    """
+    Q-target beat.
+
+    Represents a single beat in a quantization target.
+
+    ..  container:: example
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ... )
+
+        >>> abjad.f(q_target_beat)
+        abjadext.qtargetitems.QTargetBeat(
+            beatspan=abjad.Duration(1, 8),
+            offset_in_ms=abjad.Offset((1500, 1)),
+            search_tree=abjadext.searchtrees.UnweightedSearchTree(
+                definition={   3: None,
+                    },
+                ),
+            tempo=abjad.MetronomeMark(
+                reference_duration=abjad.Duration(1, 4),
+                units_per_minute=56,
+                ),
+            )
+
+    Not composer-safe.
+
+    Used internally by ``Quantizer``.
+    """
+
+    ### CLASS VARIABLES ###
+
+    __slots__ = (
+        "_beatspan",
+        "_distances",
+        "_grouping",
+        "_offset_in_ms",
+        "_q_events",
+        "_q_grid",
+        "_q_grids",
+        "_search_tree",
+        "_tempo",
+    )
+
+    _publish_storage_format = True
+
+    ### INITIALIZER ###
+
+    def __init__(self, beatspan=None, offset_in_ms=None, search_tree=None, tempo=None):
+        beatspan = beatspan or abjad.Duration(0)
+        beatspan = abjad.Duration(beatspan)
+        offset_in_ms = offset_in_ms or abjad.Duration(0)
+        offset_in_ms = abjad.Offset(offset_in_ms)
+
+        if search_tree is None:
+            search_tree = UnweightedSearchTree()
+        assert isinstance(search_tree, SearchTree)
+        tempo = tempo or abjad.MetronomeMark((1, 4), 60)
+        # tempo = abjad.MetronomeMark(tempo)
+        if isinstance(tempo, tuple):
+            tempo = abjad.MetronomeMark(*tempo)
+        assert not tempo.is_imprecise
+
+        q_events = []
+        q_grids = []
+
+        self._beatspan = beatspan
+        self._distances = {}
+        self._offset_in_ms = offset_in_ms
+        self._q_events = q_events
+        self._q_grid = None
+        self._q_grids = q_grids
+        self._search_tree = search_tree
+        self._tempo = tempo
+
+    ### SPECIAL METHODS ###
+
+    def __call__(self, job_id):
+        """
+        Calls q-target beat.
+
+        Returns quantization job.
+        """
+        if not self.q_events:
+            return None
+        assert all(isinstance(x, QEvent) for x in self.q_events)
+        q_event_proxies = []
+        for q_event in self.q_events:
+            q_event_proxy = QEventProxy(
+                q_event,
+                self.offset_in_ms,
+                self.offset_in_ms + self.duration_in_ms,
+            )
+            q_event_proxies.append(q_event_proxy)
+        return QuantizationJob(job_id, self.search_tree, q_event_proxies)
+
+    def __format__(self, format_specification="") -> str:
+        """
+        Formats q-event.
+
+        Set `format_specification` to `''` or `'storage'`. Interprets `''`
+        equal to `'storage'`.
+        """
+        if format_specification in ("", "storage"):
+            return abjad.StorageFormatManager(self).get_storage_format()
+        return str(self)
+
+    ### PUBLIC PROPERTIES ###
+
+    @property
+    def beatspan(self) -> abjad.Duration:
+        """
+        Beatspan of q-target beat.
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ... )
+
+        >>> q_target_beat.beatspan
+        Duration(1, 8)
+
+        """
+        return self._beatspan
+
+    @property
+    def distances(self) -> typing.Tuple:
+        """
+        A list of computed distances between the ``QEventProxies``
+        associated with a ``QTargetBeat`` instance, and each ``QGrid``
+        generated for that beat.
+
+        Used internally by the ``Quantizer``.
+        """
+        return self._distances
+
+    @property
+    def duration_in_ms(self) -> abjad.Duration:
+        """
+        Duration in milliseconds of the q-target beat.
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ...     )
+
+        >>> q_target_beat.duration_in_ms
+        Duration(3750, 7)
+
+        """
+        return self.tempo.duration_to_milliseconds(self.beatspan)
+
+    @property
+    def offset_in_ms(self) -> abjad.Offset:
+        """
+        Offset in milliseconds of q-target beat.
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ... )
+
+        >>> q_target_beat.offset_in_ms
+        Offset((1500, 1))
+
+        """
+        return self._offset_in_ms
+
+    @property
+    def q_events(self) -> typing.List:
+        """
+        A list for storing ``QEventProxy`` instances.
+
+        Used internally by the ``Quantizer``.
+        """
+        return self._q_events
+
+    @property
+    def q_grid(self) -> QGrid:
+        """
+        The ``QGrid`` instance selected by a ``Heuristic``.
+
+        Used internally by the ``Quantizer``.
+        """
+        return self._q_grid
+
+    @property
+    def q_grids(self) -> typing.Tuple:
+        """
+        A tuple of ``QGrids`` generated by a ``QuantizationJob``.
+
+        Used internally by the ``Quantizer``.
+        """
+        return self._q_grids
+
+    @property
+    def search_tree(self):
+        """
+        Search tree of q-target beat.
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ...     )
+
+        >>> q_target_beat.search_tree
+        UnweightedSearchTree(definition={3: None})
+
+        Returns search tree.
+        """
+        return self._search_tree
+
+    @property
+    def tempo(self) -> abjad.MetronomeMark:
+        """
+        MetronomeMark of q-target beat.
+
+        >>> beatspan = (1, 8)
+        >>> offset_in_ms = 1500
+        >>> search_tree = abjadext.nauert.UnweightedSearchTree({3: None})
+        >>> tempo = abjad.MetronomeMark((1, 4), 56)
+
+        >>> q_target_beat = abjadext.nauert.QTargetBeat(
+        ...     beatspan=beatspan,
+        ...     offset_in_ms=offset_in_ms,
+        ...     search_tree=search_tree,
+        ...     tempo=tempo,
+        ...     )
+
+        >>> q_target_beat.tempo
+        MetronomeMark(reference_duration=Duration(1, 4), units_per_minute=56)
+
+        """
+        return self._tempo
 
 
 class QTargetMeasure:
@@ -26,7 +302,7 @@ class QTargetMeasure:
         ... )
 
         >>> abjad.f(q_target_measure)
-        abjadext.nauert.QTargetMeasure(
+        abjadext.qtargetitems.QTargetMeasure(
             offset_in_ms=abjad.Offset((1000, 1)),
             search_tree=abjadext.searchtrees.UnweightedSearchTree(
                 definition={   2: None,
@@ -180,7 +456,7 @@ class QTargetMeasure:
             >>> for q_target_beat in q_target_measure.beats:
             ...     abjad.f(q_target_beat)
             ...
-            abjadext.nauert.QTargetBeat(
+            abjadext.qtargetitems.QTargetBeat(
                 beatspan=abjad.Duration(1, 4),
                 offset_in_ms=abjad.Offset((1000, 1)),
                 search_tree=abjadext.searchtrees.UnweightedSearchTree(
@@ -192,7 +468,7 @@ class QTargetMeasure:
                     units_per_minute=60,
                     ),
                 )
-            abjadext.nauert.QTargetBeat(
+            abjadext.qtargetitems.QTargetBeat(
                 beatspan=abjad.Duration(1, 4),
                 offset_in_ms=abjad.Offset((2000, 1)),
                 search_tree=abjadext.searchtrees.UnweightedSearchTree(
@@ -204,7 +480,7 @@ class QTargetMeasure:
                     units_per_minute=60,
                     ),
                 )
-            abjadext.nauert.QTargetBeat(
+            abjadext.qtargetitems.QTargetBeat(
                 beatspan=abjad.Duration(1, 4),
                 offset_in_ms=abjad.Offset((3000, 1)),
                 search_tree=abjadext.searchtrees.UnweightedSearchTree(
@@ -216,7 +492,7 @@ class QTargetMeasure:
                     units_per_minute=60,
                     ),
                 )
-            abjadext.nauert.QTargetBeat(
+            abjadext.qtargetitems.QTargetBeat(
                 beatspan=abjad.Duration(1, 4),
                 offset_in_ms=abjad.Offset((4000, 1)),
                 search_tree=abjadext.searchtrees.UnweightedSearchTree(
